Tab resize patch

+++ update.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [[ "$PWD" == "$HOME" ]]; then
+		echo >&2 "Error: in $HOME directory"
+		exit 1
+fi
+
 rm -rf pkg src *.xz *.gz
 makepkg -si --noconfirm
 cd src/st-0.8.2/
